+++ src/atm/Grid.h
@@ -673,7 +673,7 @@ public:
 	///		Assign component-wise tolerance used in the nonlinear/linear convergence checks in Arkode
 	///	</summary>
 	void AssignComponentWiseTolerances(
-		int it, int vatol_vel, int vatol_rho, int vatol_theta
+		int it, double vatol_vel, double vatol_rho, double vatol_theta
 	) const;
 
 public:
